backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymssql
python-dotenv